requirements.txt
telethon
requests
cryptg
telethon-tgcrypto
pet-pet-gif
python-dotenv